+++ subsys/bluetooth/controller/ll_sw/lll.h
@@ -263,7 +263,6 @@ enum node_rx_type {
 	NODE_RX_TYPE_SYNC_LOST,
 	NODE_RX_TYPE_SYNC_CHM_COMPLETE,
 	NODE_RX_TYPE_SYNC_ISO,
-	NODE_RX_TYPE_SYNC_ISO_PDU,
 	NODE_RX_TYPE_SYNC_ISO_LOST,
 	NODE_RX_TYPE_EXT_ADV_TERMINATE,
 	NODE_RX_TYPE_BIG_COMPLETE,
